+++ src/types/mod.rs
@@ -5,13 +5,13 @@ pub use constraints::{EdgeConstraint, GlobalConstraint, VertexConstraint, Vertex
 mod constraints;
 
 /// The vertex id type.
-pub type VId = i64;
+pub type VId = i32;
 
 /// The vertex label type.
-pub type VLabel = i64;
+pub type VLabel = i16;
 
 /// The edge label type.
-pub type ELabel = i64;
+pub type ELabel = i16;
 
 pub(crate) struct VLabelPosLen {
     pub vlabel: VLabel,
@@ -22,21 +22,21 @@ pub(crate) struct VLabelPosLen {
 pub(crate) struct VertexHeader {
     pub num_bytes: usize,
     pub vid: VId,
-    pub in_deg: usize,
-    pub out_deg: usize,
-    pub num_vlabels: usize,
+    pub in_deg: u32,
+    pub out_deg: u32,
+    pub num_vlabels: u16,
 }
 
 pub(crate) struct NeighborHeader {
     pub nid: VId,
-    pub num_n_to_v: usize,
-    pub num_v_to_n: usize,
+    pub num_n_to_v: u16,
+    pub num_v_to_n: u16,
 }
 
 pub(crate) struct SuperRowHeader {
-    pub num_rows: usize,
-    pub num_eqvs: usize,
-    pub num_vertices: usize,
+    pub num_rows: u32,
+    pub num_eqvs: u32,
+    pub num_vertices: u32,
 }
 
 pub(crate) struct PosLen {
